requirements.txt
langchain>=0.0.250
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-docx>=0.8.11
docx2txt>=0.8
llama-cpp-python>=0.1.51
tqdm
python-dotenv